requirements.txt
pydantic==2.9.2
pyyaml==6.0.2
streamlit==1.38.0
boto3==1.34.131
click==8.1.7
tabulate==0.9.0
colorama==0.4.6
python-dotenv==1.0.1
pandas==2.2.3
flake8==7.1.0